Update README.md
Infertility has become a pressing issue affecting numerous couples globally, prompting diverse approaches to address its complexities. Recent data from the National Center for Health Statistics (NCHS) indicate that infertility affects 8.8% of the American population. According to a study conducted in the United Kingdom, the prevalence of female infertility was found to be  12.5%,  while  the  prevalence  of  male  infertility  was  estimated  to  be  10.1% [1]. Extensive research  has  explored  various  diagnostic  models  in  reproductive  medicine,  reflecting  the dynamic nature of medical science and its ongoing adaptation to new ideas and challenges.